results/final_count/run-20230926-130438_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_10-runs_log/twitter/8/LLMatic/test_search.py
import pytest
from search import Search

def test_search_by_keyword():
	search = Search(['Hello world', 'Python is cool', 'I love programming'])
	assert search.search_by_keyword('Python') == ['Python is cool']


def test_filter_by_hashtag():
	search = Search(['#Hello world', 'Python is #cool', 'I love #programming'])
	assert search.filter_by_hashtag('cool') == ['Python is #cool']


def test_filter_by_user_mention():
	search = Search(['@John Hello world', 'Python is cool', 'I love programming'])
	assert search.filter_by_user_mention('John') == ['@John Hello world']


def test_filter_by_trending():
	search = Search(['#Hello world', 'Python is #cool', 'I love #programming'])
	assert search.filter_by_trending(['Hello', 'cool']) == ['#Hello world', 'Python is #cool']
